rename: api 파일 구조 변경 (controller, dto)
