+++ cgv_clone/src/main/java/com/ceos22/cgv_clone/api/controller/HelloController.java
@@ -1,7 +1,5 @@
-package com.ceos22.cgv_clone.api;
+package com.ceos22.cgv_clone.api.controller;
 
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
