+++ cgv_clone/src/main/java/com/ceos22/cgv_clone/api/controller/MovieController.java
@@ -1,7 +1,7 @@
-package com.ceos22.cgv_clone.api;
+package com.ceos22.cgv_clone.api.controller;
 
 
-import com.ceos22.cgv_clone.domain.dto.Movie;
+import com.ceos22.cgv_clone.api.dto.Movie;
 import com.ceos22.cgv_clone.service.movie.FindMovieService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
