+++ cgv_clone/src/main/java/com/ceos22/cgv_clone/api/controller/ReservationController.java
@@ -1,8 +1,8 @@
-package com.ceos22.cgv_clone.api;
+package com.ceos22.cgv_clone.api.controller;
 
-import com.ceos22.cgv_clone.domain.dto.CreateReservationCommand;
-import com.ceos22.cgv_clone.domain.dto.ReservationSummaryDto;
-import com.ceos22.cgv_clone.domain.dto.SeatSelection;
+import com.ceos22.cgv_clone.api.dto.CreateReservationCommand;
+import com.ceos22.cgv_clone.api.dto.ReservationSummaryDto;
+import com.ceos22.cgv_clone.api.dto.SeatSelection;
 import com.ceos22.cgv_clone.service.ReservationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
