+++ cgv_clone/src/main/java/com/ceos22/cgv_clone/api/dto/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.ceos22.cgv_clone.domain.dto;
+package com.ceos22.cgv_clone.api.dto;
 
 public record LoginResponse(
         String accessToken
