+++ cgv_clone/src/main/java/com/ceos22/cgv_clone/api/dto/Movie.java
@@ -1,4 +1,4 @@
-package com.ceos22.cgv_clone.domain.dto;
+package com.ceos22.cgv_clone.api.dto;
 
 import com.ceos22.cgv_clone.domain.reservationMovie.MovieEntity;
