+++ cgv_clone/src/main/java/com/ceos22/cgv_clone/api/dto/SignUpRequest.java
@@ -1,11 +1,12 @@
-package com.ceos22.cgv_clone.domain.dto;
+package com.ceos22.cgv_clone.api.dto;
 
 import com.ceos22.cgv_clone.domain.member.Gender;
 
-public record CreateMemberRequest(
+public record SignUpRequest(
         String name,
         int age,
         Gender gender,
         String loginId,
         String password
-) {}
+) {
+}
