--- cgv_clone/src/main/java/com/ceos22/cgv_clone/domain/dto/CreateFavoriteMovieCommand.java
@@ -1,6 +1,0 @@
-package com.ceos22.cgv_clone.domain.dto;
-
-public record CreateFavoriteMovieCommand(
-        Long id,
-        String movieTitle
-) {}
