+++ cgv_clone/src/main/java/com/ceos22/cgv_clone/service/movie/MovieReader.java
@@ -1,14 +1,10 @@
 package com.ceos22.cgv_clone.service.movie;
 
-import com.ceos22.cgv_clone.domain.dibsOn.FavoriteMovieEntity;
-import com.ceos22.cgv_clone.domain.dto.Movie;
+import com.ceos22.cgv_clone.api.dto.Movie;
 import com.ceos22.cgv_clone.domain.reservationMovie.MovieEntity;
-import com.ceos22.cgv_clone.repository.FavoriteMovieRepository;
 import com.ceos22.cgv_clone.repository.MovieRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -19,7 +15,7 @@ public class MovieReader {
     /** 영화 단건 조회 */
     public Movie findById(Long movieId) {
         MovieEntity movie = movieRepository.findById(movieId)
-                .orElseThrow(null);
+                .orElseThrow();
         return Movie.from(movie);
     }
 }
